+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.66.0
 - Add `walkAsync`
 - Add `findIndexedAsync` (used internally and not documented, but exported for testing)
+- Add `mapArgs`
 
 # 1.65.0
 - Add support for regular expression arguments to `highlight`